+++ src/scene/cubeMap.cu
@@ -4,7 +4,7 @@
 #include "../ui/TraceUI.h"
 #include <glm/gtx/io.hpp>
 
-glm::dvec3 CubeMap::getColor(const ray& r) const {
+glm::vec3 CubeMap::getColor(const ray& r) const {
 
     // First, map a unit direction to the cube
     auto dir = glm::normalize(r.getDirection());
@@ -14,7 +14,7 @@ glm::dvec3 CubeMap::getColor(const ray& r) const {
     if (std::abs(dir[0]) > std::abs(dir[2]) && std::abs(dir[0]) > std::abs(dir[1]))
         idx = 0;
 
-    auto forward = glm::dvec3((dir[idx] < 0 ? -1 : 1) * (idx == 0),
+    auto forward = glm::vec3((dir[idx] < 0 ? -1 : 1) * (idx == 0),
                               (dir[idx] < 0 ? -1 : 1) * (idx == 1),
                               (dir[idx] < 0 ? -1 : 1) * (idx == 2));
     dir /= glm::dot(dir, forward);
@@ -24,9 +24,9 @@ glm::dvec3 CubeMap::getColor(const ray& r) const {
     if (dir[idx / 2] < 0 != (idx == 4))
         idx++;
 
-    auto up = idx / 2 == 1 ? glm::dvec3(0, 0, (dir[idx / 2] < 0 ? -1 : 1)) : glm::dvec3(0, 1, 0);
+    auto up = idx / 2 == 1 ? glm::vec3(0, 0, (dir[idx / 2] < 0 ? -1 : 1)) : glm::vec3(0, 1, 0);
     auto right = glm::normalize(glm::cross(forward, up));
-    auto uv = glm::dvec2(0.5 + glm::dot(dir, right) / 2, 0.5 + glm::dot(dir, up) / 2);
+    auto uv = glm::dvec2(0.5f + glm::dot(dir, right) / 2, 0.5f + glm::dot(dir, up) / 2);
     return tMap[idx]->getMappedValue(uv);
 }
