+++ src/scene/material.h
@@ -36,13 +36,13 @@ public:
     // is assumed to be within the parametrization space:
     // [0, 1] x [0, 1]
     // (i.e., {(u, v): 0 <= u <= 1 and 0 <= v <= 1}
-    glm::dvec3 getMappedValue(const glm::dvec2 &coord) const;
+    glm::vec3 getMappedValue(const glm::vec2 &coord) const;
 
     // Retrieve the value stored in a physical location
     // (with integer coordinates) in the bitmap.
     // Should be called from getMappedValue in order to
     // do bilinear interpolation.
-    glm::dvec3 getPixelAt(int x, int y) const;
+    glm::vec3 getPixelAt(int x, int y) const;
 
     ~TextureMap() = default;
 
@@ -83,9 +83,9 @@ public:
 
     ~Material() = default;
 
-    Material(const glm::dvec3 &e, const glm::dvec3 &d,
-             const glm::dvec3 &r, const glm::dvec3 &t, double in)
-            : _ke(e), _kd(d), _kr(r), _kt(t), _index(glm::dvec3(in, in, in)) {}
+    Material(const glm::vec3 &e, const glm::vec3 &d,
+             const glm::vec3 &r, const glm::vec3 &t, float in)
+            : _ke(e), _kd(d), _kr(r), _kt(t), _index(glm::vec3(in, in, in)) {}
 
     __host__ __device__
     Material &
@@ -98,44 +98,44 @@ public:
         return *this;
     }
 
-    __host__ __device__ friend Material operator*(double d, Material m);
+    __host__ __device__ friend Material operator*(float d, Material m);
 
     // Accessor functions; we pass in an isect& for cases where
     // the parameter is dependent on, for example, world-space
     // coordinates (i.e., solid textures) or parametrized coordinates
     // (i.e., mapped textures)
-    __host__ __device__ glm::dvec3 ke(const isect &i) const { return _ke; }
+    __host__ __device__ glm::vec3 ke(const isect &i) const { return _ke; }
 
-    __host__ __device__ glm::dvec3 kd(const isect &i) const { return _kd; }
+    __host__ __device__ glm::vec3 kd(const isect &i) const { return _kd; }
 
-    __host__ __device__ glm::dvec3 kr(const isect &i) const { return _kr; }
+    __host__ __device__ glm::vec3 kr(const isect &i) const { return _kr; }
 
-    __host__ __device__ glm::dvec3 kt(const isect &i) const { return _kt; }
+    __host__ __device__ glm::vec3 kt(const isect &i) const { return _kt; }
 
-    __host__ __device__ double index(const isect &i) const { return _index[0]; }
+    __host__ __device__ float index(const isect &i) const { return _index[0]; }
 
-    void setEmissive(const glm::dvec3 &ke) { _ke = ke; }
+    void setEmissive(const glm::vec3 &ke) { _ke = ke; }
 
-    void setDiffuse(const glm::dvec3 &kd) { _kd = kd; }
+    void setDiffuse(const glm::vec3 &kd) { _kd = kd; }
 
-    void setReflective(const glm::dvec3 &kr) { _kr = kr; }
+    void setReflective(const glm::vec3 &kr) { _kr = kr; }
 
-    void setTransmissive(const glm::dvec3 &kt) { _kt = kt; }
+    void setTransmissive(const glm::vec3 &kt) { _kt = kt; }
 
-    void setIndex(const glm::dvec3 &index) { _index = index; }
+    void setIndex(const glm::vec3 &index) { _index = index; }
 
 private:
-    glm::dvec3 _ke = glm::dvec3(0);
-    glm::dvec3 _kd = glm::dvec3(0);
-    glm::dvec3 _kr = glm::dvec3(0);
-    glm::dvec3 _kt = glm::dvec3(0);
-    glm::dvec3 _index = glm::dvec3(1.0);
+    glm::vec3 _ke = glm::vec3(0);
+    glm::vec3 _kd = glm::vec3(0);
+    glm::vec3 _kr = glm::vec3(0);
+    glm::vec3 _kt = glm::vec3(0);
+    glm::vec3 _index = glm::vec3(1.0);
 };
 
 // This doesn't necessarily make sense for mapped materials
 __host__ __device__
 inline Material
-operator*(double d, Material m) {
+operator*(float d, Material m) {
     m._ke *= d;
     m._kd *= d;
     m._kr *= d;
